move inference providers into package
